+++ core/data-resolver/NewsletterService.ts
@@ -7,11 +7,6 @@ import getApiEndpointUrl from '@vue-storefront/core/helpers/getApiEndpointUrl';
 const isSubscribed = (email: string): Promise<boolean> =>
   TaskQueue.execute({
     url: processURLAddress(getApiEndpointUrl(config.newsletter, 'endpoint')) + '?email=' + encodeURIComponent(email),
-    payload: {
-      method: 'GET',
-      headers: { 'Content-Type': 'application/json' },
-      mode: 'cors'
-    },
     silent: true
   }).then(({ result }) => result === 'subscribed')
 
@@ -20,8 +15,6 @@ const subscribe = (email: string): Promise<boolean> =>
     url: processURLAddress(getApiEndpointUrl(config.newsletter, 'endpoint')),
     payload: {
       method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      mode: 'cors',
       body: JSON.stringify({ email })
     }
   }).then(({ code }) => code === 200)
@@ -31,8 +24,6 @@ const unsubscribe = (email: string): Promise<boolean> =>
     url: processURLAddress(getApiEndpointUrl(config.newsletter, 'endpoint')),
     payload: {
       method: 'DELETE',
-      headers: { 'Content-Type': 'application/json' },
-      mode: 'cors',
       body: JSON.stringify({ email })
     }
   }).then(({ code }) => code === 200)