+++ core/data-resolver/ProductService.ts
@@ -110,14 +110,6 @@ const getProductRenderList = async ({
 
   try {
     const task = await TaskQueue.execute({ url, // sync the cart
-      payload: {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-          ...(token && config.users.tokenInHeader ? { authorization: `Bearer ${token}` } : {})
-        },
-        mode: 'cors'
-      },
       callback_event: 'prices-after-sync'
     })
     return task.result as DataResolver.ProductsListResponse