+++ core/lib/async-data-loader.ts
@@ -2,8 +2,8 @@
 * @deprecated This module has been created before Vue.js v2.6.0. From 2.6.x on please do preferably use the `serverPrefetch` hook to sync SSR Vuex data. More info: https://ssr.vuejs.org/api/#serverprefetch
 * @description Please note: This module can be used not only for `asyncData` extensibility but also for simplyfying the data loaders in Vuex actions or other components - basicaly everywhere where we must `wait` for some async operations to complete.
 */
-import { isServer } from '@vue-storefront/core/helpers'
 import { Logger } from '@vue-storefront/core/lib/logger'
+import { Context } from './../scripts/utils/types'
 
 const DEFAULT_ACTION_CATEGORY = 'asyncData'
 // Data loader queues all the data fetching operations and runs them at once - to be usedf for example in the `asyncData()` functions
@@ -11,12 +11,12 @@ export interface AsyncDataLoaderActionContext {
   category?: string,
   route: any,
   store: any,
-  context: any
+  context: Context
 }
 
 // Data loader queues all the data fetching operations and runs them at once - to be usedf for example in the `asyncData()` functions
 export interface AsyncDataLoaderAction {
-  execute: any, // this function must return a Promise
+  execute: (args: AsyncDataLoaderActionContext) => Promise<void>, // this function must return a Promise
   category?: string,
   name?: string,
   executedAt?: Date,