+++ core/lib/sync/task.ts
@@ -51,10 +51,14 @@ function getUrl (task, currentToken, currentCartId) {
 
 function getPayload (task, currentToken) {
   const payload = {
-    ...task.payload,
+    mode: 'cors',
+    method: 'GET',
+    ...(task?.payload || {}),
     headers: {
-      ...task.payload.headers,
-      ...(config.users.tokenInHeader ? { authorization: `Bearer ${currentToken}` } : {})
+      'Content-Type': 'application/json',
+      'Accept': 'application/json, text/plain, */*',
+      ...(task?.payload?.headers || {}),
+      ...(config.users.tokenInHeader ? { [config.users.tokenHeaderName]: `${config.users.tokenAuthScheme} ${currentToken}` } : {})
     }
   }
   return payload