+++ core/modules/catalog-next/store/category/actions.ts
@@ -24,7 +24,7 @@ import createCategoryListQuery from '@vue-storefront/core/modules/catalog/helper
 import { transformCategoryUrl } from '@vue-storefront/core/modules/url/helpers/transformUrl';
 
 const actions: ActionTree<CategoryState, RootState> = {
-  async loadCategoryProducts ({ commit, getters, dispatch, rootState }, { route, category, pageSize = 50 } = {}) {
+  async loadCategoryProducts ({ commit, getters, dispatch }, { route, category, pageSize = 50 } = {}) {
     const searchCategory = category || getters.getCategoryFrom(route.path) || {}
     const categoryMappedFilters = getters.getFiltersMap[searchCategory.id]
     const areFiltersInQuery = !!Object.keys(route[products.routerFiltersSource]).length
@@ -60,7 +60,7 @@ const actions: ActionTree<CategoryState, RootState> = {
 
     return items
   },
-  async loadMoreCategoryProducts ({ commit, getters, rootState, dispatch }) {
+  async loadMoreCategoryProducts ({ commit, getters, dispatch }) {
     const { perPage, start, total } = getters.getCategorySearchProductsStats
     const totalValue = typeof total === 'object' ? total.value : total
     if (start >= totalValue || totalValue < perPage) return
@@ -94,7 +94,7 @@ const actions: ActionTree<CategoryState, RootState> = {
 
     return searchResult.items
   },
-  async cacheProducts ({ commit, getters, dispatch, rootState }, { route } = {}) {
+  async cacheProducts ({ getters, dispatch, rootState }, { route } = {}) {
     if (config.api.saveBandwidthOverCache) {
       return
     }
@@ -119,7 +119,7 @@ const actions: ActionTree<CategoryState, RootState> = {
       }
     }
   },
-  async findCategories (context, categorySearchOptions: DataResolver.CategorySearchOptions): Promise<Category[]> {
+  async findCategories (_, categorySearchOptions: DataResolver.CategorySearchOptions): Promise<Category[]> {
     return CategoryService.getCategories(categorySearchOptions)
   },
   async loadCategories ({ commit, getters }, categorySearchOptions: DataResolver.CategorySearchOptions): Promise<Category[]> {
@@ -260,9 +260,9 @@ const actions: ActionTree<CategoryState, RootState> = {
         }, { root: true })
       }
     }
-  },
+  }
   /** Below actions are not used from 1.12 and can be removed to reduce bundle */
-  ...require('./deprecatedActions').default
+  // ...require('./deprecatedActions').default
 }
 
 export default actions