+++ core/modules/catalog/helpers/associatedProducts/index.ts
@@ -1,9 +1,5 @@
-import setGroupedProduct from './setGroupedProduct'
-import setBundleProducts from './setBundleProducts'
-import getAttributesFromMetadata from './getAttributesFromMetadata'
+import getAndCallAsyncDefault from '@vue-storefront/core/helpers/getAndCallAsyncDefault';
 
-export {
-  setGroupedProduct,
-  setBundleProducts,
-  getAttributesFromMetadata
-}
+export const setGroupedProduct = getAndCallAsyncDefault(() => import(/* webpackChunkName: "vsf-catalog-helper-pack-setLinkableProducts" */ './setGroupedProduct'));
+export const setBundleProducts = getAndCallAsyncDefault(() => import(/* webpackChunkName: "vsf-catalog-helper-pack-setLinkableProducts" */ './setBundleProducts'));
+export const getAttributesFromMetadata = getAndCallAsyncDefault(() => import(/* webpackChunkName: "vsf-catalog-helper-getAttributesFromMetadata" */ './getAttributesFromMetadata'));