+++ core/modules/url/test/unit/helpers/formatCategoryLink.spec.ts
@@ -1,6 +1,6 @@
 import { formatCategoryLink } from '@vue-storefront/core/modules/url/helpers';
 import { Category } from '@vue-storefront/core/modules/catalog-next/types/Category';
-import { currentStoreView } from '@vue-storefront/core/lib/multistore';
+import { currentStoreView, localizedRoute } from '@vue-storefront/core/lib/multistore';
 import config from 'config';
 
 jest.mock('@vue-storefront/core/app', () => jest.fn());
@@ -21,6 +21,7 @@ describe('formatCategoryLink method', () => {
     jest.clearAllMocks();
     jest.mock('config', () => ({}));
     (currentStoreView as jest.Mock).mockImplementation(() => ({ storeCode: '' }));
+    (localizedRoute as jest.Mock).mockImplementation(() => '');
     category = {
       path: '1/2',
       is_active: true,
@@ -45,44 +46,19 @@ describe('formatCategoryLink method', () => {
     });
 
     it('should return formatted category url_path', () => {
-      const result = formatCategoryLink(category);
-      expect(result).toEqual('/all-2/women/women-20');
-    });
-
-    it('should return formatted category url_path when storeCode passed as null', () => {
-      const result = formatCategoryLink(category, null);
-      expect(result).toEqual('/all-2/women/women-20');
-    });
-
-    it('should return formatted category url_path when storeCode passed as \'de\'', () => {
       const result = formatCategoryLink(category, 'de');
-      expect(result).toEqual('/de/all-2/women/women-20');
+      expect(localizedRoute).toBeCalledWith(category.url_path, 'de');
     });
 
-    it('should return formatted category url_path when storeCode passed as \'\'', () => {
-      const result = formatCategoryLink(category, '');
-      expect(result).toEqual('/all-2/women/women-20');
+    it('should return formatted homepage path when category is missing', () => {
+      const result = formatCategoryLink(null, 'de');
+      expect(localizedRoute).toBeCalledWith('/', 'de');
     });
 
-    it('should return homepage path when category passed as \'null\'', () => {
-      const result = formatCategoryLink(null);
-      expect(result).toEqual('/');
-    });
-
-    describe('with default storeCode set to \'de\'', () => {
-      beforeEach(() => {
-        (currentStoreView as jest.Mock).mockImplementation(() => ({ storeCode: 'de' }));
-      });
-
-      it('should return formatted category url_path', () => {
-        const result = formatCategoryLink(category);
-        expect(result).toEqual('/de/all-2/women/women-20');
-      });
-
-      it('should return homepage path when category passed as \'null\'', () => {
-        const result = formatCategoryLink(null);
-        expect(result).toEqual('/de/');
-      });
+    it('should use current storeCode if no storeCode is passed', () => {
+      (currentStoreView as jest.Mock).mockImplementation(() => ({ storeCode: 'de' }));
+      const result = formatCategoryLink(category);
+      expect(localizedRoute).toBeCalledWith(category.url_path, 'de');
     });
   });
 
@@ -93,56 +69,20 @@ describe('formatCategoryLink method', () => {
       };
     });
 
-    it('should return old path with c and category slug', () => {
-      const result = formatCategoryLink(category);
-      expect(result).toEqual('/c/all-2');
-    });
-
-    it('should return old path with c and category slug when storeCode passed as null', () => {
-      const result = formatCategoryLink(category, null);
-      expect(result).toEqual('/c/all-2');
-    });
-
-    it('should return old path with c and category slug when storeCode passed as \'de\'', () => {
+    it('should return formatted category slug with prefix', () => {
       const result = formatCategoryLink(category, 'de');
-      expect(result).toEqual('/de/c/all-2');
+      expect(localizedRoute).toBeCalledWith(`c/${category.slug}`, 'de');
     });
 
-    it('should return old path with c and category slug when storeCode passed as \'\'', () => {
-      const result = formatCategoryLink(category, '');
-      expect(result).toEqual('/c/all-2');
+    it('should return formatted homepage path when category is missing', () => {
+      const result = formatCategoryLink(null, 'de');
+      expect(localizedRoute).toBeCalledWith('/', 'de');
     });
 
-    describe('with default storeCode set to \'de\'', () => {
-      beforeEach(() => {
-        (currentStoreView as jest.Mock).mockImplementation(() => ({ storeCode: 'de' }));
-      });
-
-      it('should return formatted category url_path', () => {
-        const result = formatCategoryLink(category);
-        expect(result).toEqual('/de/c/all-2');
-      });
-
-      it('should return homepage path when category passed as \'null\'', () => {
-        const result = formatCategoryLink(null);
-        expect(result).toEqual('/de/');
-      });
-    })
-
-    describe('with default storeCode set to \'de\' and appendStoreCode is false', () => {
-      beforeEach(() => {
-        (currentStoreView as jest.Mock).mockImplementation(() => ({ storeCode: 'de', appendStoreCode: false }));
-      });
-
-      it('should return formatted category url_path', () => {
-        const result = formatCategoryLink(category);
-        expect(result).toEqual('/c/all-2');
-      });
-
-      it('should return homepage path when category passed as \'null\'', () => {
-        const result = formatCategoryLink(null);
-        expect(result).toEqual('/');
-      });
-    })
+    it('should use current storeCode if no storeCode is passed', () => {
+      (currentStoreView as jest.Mock).mockImplementation(() => ({ storeCode: 'de' }));
+      const result = formatCategoryLink(category);
+      expect(localizedRoute).toBeCalledWith(`c/${category.slug}`, 'de');
+    });
   });
 });