+++ core/scripts/utils/ssr-renderer.ts
@@ -1,4 +1,5 @@
 import { Context } from './types';
+import { createRedirectProxy, addRedirectTempObject } from './ssr-redirect';
 const fs = require('fs')
 const path = require('path')
 const compile = require('lodash.template')
@@ -107,7 +108,7 @@ function initTemplatesCache (config, compileOptions) {
 }
 
 function initSSRRequestContext (app, req, res, config): Context {
-  return {
+  const context: Context = {
     url: decodeURI(req.url),
     output: {
       prepend: (context) => { return ''; },
@@ -119,7 +120,7 @@ function initSSRRequestContext (app, req, res, config): Context {
     },
     server: {
       app: app,
-      response: res,
+      response: null, // we will initialize it as proxy
       request: req
     },
     meta: null,
@@ -128,6 +129,14 @@ function initSSRRequestContext (app, req, res, config): Context {
       storeCode: typeof req.header === 'function' ? (req.header('x-vs-store-code') ? req.header('x-vs-store-code') : process.env.STORE_CODE) : process.env.STORE_CODE
     }
   };
+
+  // build temporary object that will be used to resolve redirection
+  addRedirectTempObject(context, res)
+
+  // assign proxy instead of original response object
+  context.server.response = createRedirectProxy(context, res)
+
+  return context
 }
 
 function clearContext (context) {